Splits plotting rules to maps and non-maps. Adds plots. Renames policy constraints to GlobalConstraints to track in assigning duals
