+++ workflow/config/config.default.yaml
@@ -11,13 +11,13 @@ run:
 # docs :
 scenario:
   interconnect: western #"usa|texas|western|eastern"
-  clusters: [100]
+  clusters: [300]
   opts: [RCo2L-SAFER-RPS]
   ll: [v1.0]
   scope: "total" # "urban", "rural", or "total"
   sector: "" # G
   planning_horizons:
-  - 2030   #(2030, 2040, 2050)
+  # - 2030   #(2030, 2040, 2050)
 
 foresight:  # Only Single Stage Currently
 
@@ -256,7 +256,7 @@ clustering:
     resolution_sector: false
 
 focus_weights:
-  # California: 0.5
+  California: 0.5
 
 
 # docs :
